prompt for sizing, adjusts divs to fit container
this was hard, thought about and researched for a few hours and couldn't find anything. i ended up looking on to another project and found '${}' which I was unaware of/forgotten how it worked. i was able to get an expression in though like i wanted and auto-adjust the size of each etchDiv with it.